backend/requirements.txt
fastapi[all]==0.95.1
pymongo==4.3.3
pydantic==1.10.7
uvicorn==0.22.0
python-dotenv==0.20.0
